+++ tests/test_misc.py
@@ -3332,16 +3332,22 @@ def test_restorefrompeer(node_factory, bitcoind):
     l1.start()
     assert l1.daemon.is_in_log('Server started with public key')
 
-    # If this happens fast enough, connect fails with "disconnected
-    # during connection"
-    try:
-        l1.rpc.connect(l2.info['id'], 'localhost', l2.port)
-    except RpcError as err:
-        assert "disconnected during connection" in err.error['message']
+    l1.rpc.connect(l2.info['id'], 'localhost', l2.port)
 
     l1.daemon.wait_for_log('peer_in WIRE_PEER_STORAGE_RETRIEVAL')
 
+    # We lost our db, so l2's channel_reestablish is for a channel we don't
+    # know: we answer with an error, but we don't hang up on them any more.
+    l1.daemon.wait_for_log('Unknown channel .* for WIRE_CHANNEL_REESTABLISH')
+    assert only_one(l1.rpc.listpeers()['peers'])['connected']
+
     assert l1.rpc.restorefrompeer()['stubs'][0] == _['channel_id']
+
+    # We need to reconnect so the stub channel triggers the bogus
+    # channel_reestablish flow in peer_connected_hook_final.
+    # (l1 no longer disconnects on unknown channel_reestablish.)
+    l1.rpc.disconnect(l2.info['id'], force=True)
+    l1.rpc.connect(l2.info['id'], 'localhost', l2.port)
 
     l1.daemon.wait_for_log('Sending a bogus channel_reestablish message to make the peer unilaterally close the channel.')
     l1.daemon.wait_for_log('peer_out WIRE_ERROR')